requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
langchain==0.0.335
langchain-openai==0.0.2
langchain-community==0.0.10
uvicorn==0.24.0
python-dotenv==1.0.0
jinja2==3.1.2